Pull control message code out into a separate file.
